requirements.txt
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
numpy==1.22.4
astropy==5.0.4
astroplan==0.8
pytz==2023.3
requests==2.31.0
PyYAML==6.0.1 